Monster TakeDamage 수정
이번에만 TeamWork에서 작업하고 제 브런치로 돌아가겠습니다

+++ 9TeamTd/Assets/Scripts/Json/MonsterStats.cs
@@ -12,7 +12,7 @@ public class MonsterStats : MonoBehaviour
     public string name;
     public int level;
     public int maxHP;
-    public int attackValue;
+    public float attackValue;
     public float defenceValue;
     public string moveType;
     public string enemyRank;
